User/address optimize 2.13 (#19)
* fix(apps/auth):添加redis中检测失效令牌的功能

* feat(pkg/util)AliyunOss文件存储上传工具类

* feat(apps/user)user模块开发

* feat(apps/user)user模块完善(可用版本),新增自定义错误,为AliOss对象存储增加了错误处理方法

* feat(apps/address)address模块开发

* style(apps/address  apps/user)address和user模块性能优化

+++ pkg/models/user.go
@@ -9,6 +9,6 @@ type User struct {
 	Name   string `gorm:"type:varchar(60) comment '用户名称'"`
 	Email  string `gorm:"unique;comment '用户邮箱'"`
 	Phone  string `gorm:"type:varchar(11) unique comment '用户手机号'"`
-	Gender uint8  `gorm:"comment '用户性别 1男 0女'"`
+	Gender uint32 `gorm:"comment '用户性别 1男 0女'"`
 	Avatar string `gorm:"type varchar(200) comment '用户头像'"`
 }
